Updated to match NXT files
git-svn-id: http://svn.code.sf.net/p/lejos/code/trunk@3588 06c8f799-c779-4783-8af7-179afedb67c8

+++ pccomms/lejos/robotics/Pose.java
@@ -110,7 +110,7 @@ public void arcUpdate(float distance, float turnAngle)
  */
 public float angleTo(Point destination)
 {
-  return (float)_location.angleTo(destination);
+  return _location.angleTo(destination);
 }
 /**
  * Returns the angle to <code>destination</code> relative to the pose heading;
